Priority Queue Shortest Job First Operation
Written in C, this program simulates an operating system that has 3
priority queues and of the 3, the process that is expected to run in the
shortest amount of time is chosen to run first all the while new
processes are being added during runtime. All the proceses that run in
this program is in the a Processes.txt file.